pass context to controller
